Starting to save

+++ gui/src/hkx/ProjectBuilder.h
@@ -22,7 +22,7 @@ namespace ckcmd {
 			{
 				auto& content = _resourceManager.get(path);
 				int data_index = -1;
-				for (int v = 0; v < content.second.size(); v++)
+				for (size_t v = 0; v < content.second.size(); v++)
 				{
 					if (content.second[v].m_class == &hk_class)
 					{
